+++ src/ui/components/Library/Preview/Preview.tsx
@@ -1,13 +1,13 @@
 import { useContext } from "react";
 import { LibraryContext } from "../useFilters";
-import { TestPreview } from "./TestPreview";
-import { TestRunPreview } from "./TestRunPreview";
+import { TestPreview } from "./TestPreview/TestPreview";
+import { TestRunPreview } from "./TestRunPreview/TestRunPreview";
 
 export function Preview() {
   const { preview } = useContext(LibraryContext);
   return (
     <div
-      className="flex flex-col bg-white rounded-md shadow-md text-sm p-4 space-y-4 overflow-auto"
+      className="flex flex-col p-4 space-y-4 overflow-auto text-sm bg-white rounded-md shadow-md"
       style={{ width: "50rem" }}
     >
       {preview?.view === "tests" ? <TestPreview /> : <TestRunPreview />}
